requirements.txt
praw==7.7.1
python-dotenv==1.0.1
google-generativeai==0.3.2
tqdm==4.66.4
